Here's a Python regex implementation.
It's not particularly bright, but it can be thrown at the MIT
regexp crossword.

The trickiest line is the "OREO" one. Bizarrely, this is a simple
line for a human to resolve, but a right pain in the arse for the
solver. I approach it via a brute-force expansion of the expression
(which *is*, technically speaking, regular - but only because it's
anchored at a fixed length).